update depot apis
