+++ packages/legend-application-studio/src/components/editor/__test-utils__/EditorComponentTestUtils.tsx
@@ -52,7 +52,7 @@ import {
 } from '@finos/legend-graph';
 import {
   DepotServerClient,
-  type ProjectData,
+  type StoreProjectData,
   type ProjectVersionEntities,
   type RawProjectDependencyReport,
 } from '@finos/legend-server-depot';
