+++ packages/legend-extension-dsl-service/src/components/studio/UpdateServiceQuerySetup.tsx
@@ -34,7 +34,7 @@ import {
   useLegendStudioBaseStore,
 } from '@finos/legend-application-studio';
 import { UpdateServiceQuerySetupStore } from '../../stores/studio/UpdateServiceQuerySetupStore.js';
-import { type ProjectData } from '@finos/legend-server-depot';
+import type { StoreProjectData } from '@finos/legend-server-depot';
 import {
   type ServiceQueryUpdaterSetupPathParams,
   generateServiceQueryUpdaterRoute,
